Fixed docstring and avoid copies when using astype if not needed

+++ astropy/io/fits/_tiled_compression/tiled_compression.py
@@ -500,7 +500,7 @@ def compress_hdu(hdu):
                 if array.dtype.byteorder == "<" or (
                     array.dtype.byteorder == "=" and sys.byteorder == "little"
                 ):
-                    compressed_bytes[irow] = array.astype(">i2").tobytes()
+                    compressed_bytes[irow] = array.astype(">i2", copy=False).tobytes()
 
     compressed_bytes = b"".join(compressed_bytes)
